+++ packages/apis/example/lib/services/handlers/orders_handlers/transaction_handlers/retrieves_transaction_list_handler.dart
@@ -1,10 +1,10 @@
 import 'package:apis/apis.dart';
-import 'package:apis/network/remote/orders/transaction/abstract/transaction.dart';
+import 'package:apis/network/remote/orders/transaction/abstract/transaction_service.dart';
 import 'package:example/services/api_request_handler.dart';
 import 'package:example/services/api_service_registry.dart';
 import 'package:get_it/get_it.dart';
 
-class GetTransactionSingleHandler implements ApiRequestHandler {
+class GetTransactionListHandler implements ApiRequestHandler {
   @override
   Future<Map<String, dynamic>> handleRequest(
       String method, Map<String, String> params) async {
@@ -13,27 +13,24 @@ class GetTransactionSingleHandler implements ApiRequestHandler {
         final String apiVersion =
             params['api_version'] ?? ApiNetwork.apiVersion;
         final String? orderId = params['order_id'];
-        final String? transactionId = params['transaction_id'];
+        final String? sinceId = params['since_id'];
         final String? fields = params['fields'];
         final String? inShopCurrency = params['in_shop_currency'];
 
-        if (orderId == null ||
-            orderId.isEmpty ||
-            transactionId == null ||
-            transactionId.isEmpty) {
+        if (orderId == null || orderId.isEmpty) {
           return {
             "status": "error",
-            "message": "Order ID and Transaction ID are required",
+            "message": "Order ID is required",
             "timestamp": DateTime.now().toIso8601String(),
           };
         }
 
         try {
           final response =
-              await GetIt.I.get<TransactionService>().getTransactionSingle(
+              await GetIt.I.get<TransactionService>().getTransactionList(
                     apiVersion: apiVersion,
                     orderId: orderId,
-                    transactionId: transactionId,
+                    since_id: sinceId,
                     fields: fields,
                     in_shop_currency: inShopCurrency,
                   );
@@ -41,15 +38,14 @@ class GetTransactionSingleHandler implements ApiRequestHandler {
         } catch (e) {
           return {
             "status": "error",
-            "message": "Failed to get transaction: ${e.toString()}",
+            "message": "Failed to get transaction list: ${e.toString()}",
             "timestamp": DateTime.now().toIso8601String(),
           };
         }
 
       default:
         return {
-          "error":
-              "Method $method not supported for Get Transaction Single API",
+          "error": "Method $method not supported for Get Transaction List API",
         };
     }
   }
@@ -62,9 +58,9 @@ class GetTransactionSingleHandler implements ApiRequestHandler {
         'GET': [
           const ApiField(name: 'order_id', label: 'Order ID', hint: 'Order ID'),
           const ApiField(
-              name: 'transaction_id',
-              label: 'Transaction ID',
-              hint: 'Transaction ID'),
+              name: 'since_id',
+              label: 'Since ID',
+              hint: '(Optional) Return results after this ID'),
           const ApiField(
               name: 'fields',
               label: 'Fields',
