+++ packages/apis/lib/network/remote/orders/transaction/api/api_transaction_service.dart
@@ -1,6 +1,6 @@
 import 'package:apis/apis.dart';
 import 'package:apis/dio_config/api_dio_client.dart';
-import 'package:apis/network/remote/orders/transaction/abstract/transaction.dart';
+import 'package:apis/network/remote/orders/transaction/abstract/transaction_service.dart';
 import 'package:apis/network/remote/orders/transaction/freezed_model/request/create_transaction_request.dart';
 import 'package:apis/network/remote/orders/transaction/freezed_model/response/create_transaction_response.dart';
 import 'package:apis/network/remote/orders/transaction/freezed_model/response/get_transaction_count_response.dart';
